Created LabeledLineV from some of the MeasurementEditorV
